fix bug with info "The stream sample position formatter error"

+++ Box/ApertureDimensionsBox.cpp
@@ -4,14 +4,14 @@
 
 #include "ApertureDimensionsBox.h"
 
-ApertureDimensionsBox::ApertureDimensionsBox(uint32_t size, uint32_t type, uint64_t start_pos)
+ApertureDimensionsBox::ApertureDimensionsBox(uint32_t size, uint32_t type, uint32_t start_pos)
         : FullBox(size, type, start_pos) {}
 
 ApertureDimensionsBox::~ApertureDimensionsBox() {
 
 }
 
-void ApertureDimensionsBox::Parse(struct Parser *parser, uint64_t start_pos) {
+void ApertureDimensionsBox::Parse(struct Parser *parser, uint32_t start_pos) {
     FullBox::Parse(parser, start_pos);
 
     width = parser->file_reader->Read32();
