+++ Box/Box.h
@@ -17,18 +17,18 @@ public:
     uint32_t size;
     uint32_t type;
     std::string name;
-    uint64_t start_pos;
+    uint32_t start_pos;
 
     Box *child_head;
     Box *child_trail;
     Box *next;
 
 public:
-    Box(uint32_t size, uint32_t type, uint64_t start_pos);
+    Box(uint32_t size, uint32_t type, uint32_t start_pos);
     virtual ~Box();
 
 public:
-    virtual void Parse(class Parser *parser, uint64_t start_pos);
+    virtual void Parse(class Parser *parser, uint32_t start_pos);
     virtual std::string Description();
 
 public:
