+++ Box/ChunkOffsetBox.cpp
@@ -4,13 +4,13 @@
 
 #include "ChunkOffsetBox.h"
 
-ChunkOffsetBox::ChunkOffsetBox(uint32_t size, uint32_t type, uint64_t start_pos) : FullBox(size, type, start_pos) {}
+ChunkOffsetBox::ChunkOffsetBox(uint32_t size, uint32_t type, uint32_t start_pos) : FullBox(size, type, start_pos) {}
 
 ChunkOffsetBox::~ChunkOffsetBox() {
 
 }
 
-void ChunkOffsetBox::Parse(struct Parser *parser, uint64_t start_pos) {
+void ChunkOffsetBox::Parse(struct Parser *parser, uint32_t start_pos) {
     FullBox::Parse(parser, start_pos);
 
     number_of_entries = parser->file_reader->Read32();
