+++ Box/CompositionTimeToSampleBox.h
@@ -13,9 +13,9 @@ public:
     std::vector<CompositionOffsetItem> composition_offset_item_table;
 
 public:
-    CompositionTimeToSampleBox(uint32_t size, uint32_t type, uint64_t start_pos);
+    CompositionTimeToSampleBox(uint32_t size, uint32_t type, uint32_t start_pos);
     virtual ~CompositionTimeToSampleBox();
-    void Parse(struct Parser *parser, uint64_t start_pos) override;
+    void Parse(struct Parser *parser, uint32_t start_pos) override;
 };
