+++ Box/CompositionToDecodeBox.h
@@ -18,9 +18,9 @@ public:
     std::string display_end_time_formatter;
 
 public:
-    CompositionToDecodeBox(uint32_t size, uint32_t type, uint64_t start_pos);
+    CompositionToDecodeBox(uint32_t size, uint32_t type, uint32_t start_pos);
     virtual ~CompositionToDecodeBox();
-    void Parse(struct Parser *parser, uint64_t start_pos) override;
+    void Parse(struct Parser *parser, uint32_t start_pos) override;
 };
