+++ Box/DataRefBox.cpp
@@ -4,13 +4,13 @@
 
 #include "DataRefBox.h"
 
-DataRefBox::DataRefBox(uint32_t size, uint32_t type, uint64_t start_pos) : FullBox(size, type, start_pos) {}
+DataRefBox::DataRefBox(uint32_t size, uint32_t type, uint32_t start_pos) : FullBox(size, type, start_pos) {}
 
 DataRefBox::~DataRefBox() {
 
 }
 
-void DataRefBox::Parse(struct Parser *parser, uint64_t start_pos) {
+void DataRefBox::Parse(struct Parser *parser, uint32_t start_pos) {
     FullBox::Parse(parser, start_pos);
 
     FileReader *file_reader = parser->file_reader;
