+++ Box/DataRefBox.h
@@ -22,9 +22,9 @@ public:
     std::vector<DataReferenceItem> data_reference_item_table;
 
 public:
-    DataRefBox(uint32_t size, uint32_t type, uint64_t start_pos);
+    DataRefBox(uint32_t size, uint32_t type, uint32_t start_pos);
     virtual ~DataRefBox();
-    void Parse(struct Parser *parser, uint64_t start_pos) override;
+    void Parse(struct Parser *parser, uint32_t start_pos) override;
 };
