+++ Box/FileTypeBox.cpp
@@ -4,14 +4,14 @@
 
 #include "FileTypeBox.h"
 
-FileTypeBox::FileTypeBox(uint32_t size, uint32_t type, uint64_t start_pos) : Box(size, type, start_pos) {
+FileTypeBox::FileTypeBox(uint32_t size, uint32_t type, uint32_t start_pos) : Box(size, type, start_pos) {
 }
 
 FileTypeBox::~FileTypeBox() {
 
 }
 
-void FileTypeBox::Parse(struct Parser *parser, uint64_t start_pos) {
+void FileTypeBox::Parse(struct Parser *parser, uint32_t start_pos) {
     FileReader *file_reader = parser->file_reader;
     file_reader->Seek(start_pos);
