+++ Box/FreeSpaceBox.h
@@ -9,9 +9,9 @@
 
 class FreeSpaceBox : public Box{
 public:
-    FreeSpaceBox(uint32_t size, uint32_t type, uint64_t start_pos);
+    FreeSpaceBox(uint32_t size, uint32_t type, uint32_t start_pos);
     virtual ~FreeSpaceBox();
-    virtual void Parse(struct Parser *parser, uint64_t start_pos) override;
+    virtual void Parse(struct Parser *parser, uint32_t start_pos) override;
 
 public:
     std::vector<uint8_t> data;
