+++ Box/FullBox.cpp
@@ -4,7 +4,7 @@
 
 #include "FullBox.h"
 
-FullBox::FullBox(uint32_t size, uint32_t type, uint64_t start_pos) : Box(size, type, start_pos), version(0), flags(0) {
+FullBox::FullBox(uint32_t size, uint32_t type, uint32_t start_pos) : Box(size, type, start_pos), version(0), flags(0) {
 
 }
 
@@ -12,7 +12,7 @@ FullBox::~FullBox() {
 
 }
 
-void FullBox::Parse(struct Parser *parser, uint64_t start_pos) {
+void FullBox::Parse(struct Parser *parser, uint32_t start_pos) {
     FileReader *fileReader = parser->file_reader;
     fileReader->Seek(start_pos);
