+++ Box/SampleToChunkBox.cpp
@@ -4,12 +4,12 @@
 
 #include "SampleToChunkBox.h"
 
-SampleToChunkBox::SampleToChunkBox(uint32_t size, uint32_t type, uint64_t start_pos) : FullBox(size, type, start_pos) {}
+SampleToChunkBox::SampleToChunkBox(uint32_t size, uint32_t type, uint32_t start_pos) : FullBox(size, type, start_pos) {}
 
 SampleToChunkBox::~SampleToChunkBox() {
 }
 
-void SampleToChunkBox::Parse(struct Parser *parser, uint64_t start_pos) {
+void SampleToChunkBox::Parse(struct Parser *parser, uint32_t start_pos) {
     FullBox::Parse(parser, start_pos);
 
     FileReader *file_reader = parser->file_reader;
