+++ Box/SoundMediaHeaderBox.h
@@ -12,11 +12,11 @@ public:
     uint16_t balance;
 
 public:
-    SoundMediaHeaderBox(uint32_t size, uint32_t type, uint64_t start_pos);
+    SoundMediaHeaderBox(uint32_t size, uint32_t type, uint32_t start_pos);
 
     virtual ~SoundMediaHeaderBox();
 
-    void Parse(struct Parser *parser, uint64_t start_pos) override;
+    void Parse(struct Parser *parser, uint32_t start_pos) override;
 };
