+++ Box/TrackHeaderBox.h
@@ -27,9 +27,9 @@ public:
     uint32_t track_height;
 
 public:
-    TrackHeaderBox(uint32_t size, uint32_t type, uint64_t start_pos);
+    TrackHeaderBox(uint32_t size, uint32_t type, uint32_t start_pos);
     virtual ~TrackHeaderBox();
-    void Parse(struct Parser *parser, uint64_t start_pos) override;
+    void Parse(struct Parser *parser, uint32_t start_pos) override;
 };
